# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Shopping API for Revel.",
        "title": "Revel Shopping API",
        "contact": {
            "name": "Simas Paulikas",
            "email": "simaspaulikas@yahoo.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4141",
    "basePath": "/api/v1",
    "paths": {
        "/carts/{customer_id}": {
            "get": {
                "description": "Removes item from catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Removes item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the item to retrieve",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.CartItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts/{customer_id}/items/{item_id}": {
            "post": {
                "description": "Removes item from catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Removes item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the customer to whose cart the item is added",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of an item that is being added",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.CartItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts/{customer_id}/orders": {
            "post": {
                "description": "Removes item from catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Removes item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the customer who is making the order",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order has been place",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.CartItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/carts/{customer_id}/totals": {
            "post": {
                "description": "Removes item from catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Removes item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the customer for whom to retrieve total",
                        "name": "customer_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/model.CartItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/items/{item_id}": {
            "get": {
                "description": "get item by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Show an item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the item to retrieve",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Item"
                        }
                    },
                    "400": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create an item",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Item"
                        }
                    },
                    "400": {
                        "description": "Item already exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes item from catalog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Removes item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the item to delete",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Item"
                        }
                    },
                    "400": {
                        "description": "Item doesn't exist",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CartItem": {
            "type": "object",
            "properties": {
                "quantity": {
                    "type": "integer"
                },
                "unit_price": {
                    "type": "integer"
                }
            }
        },
        "model.Item": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        }
    }
}